Rename the getPassiveVoice[Language]Spec to getPassiveVoiceResult[Language]Spec
